refactor(blocks): move SwatchbookProvider and theme hooks from addon to blocks
Flips the workspace dep graph so it runs addon → blocks (the direction
it was always meant to). Provider, contexts, and theme hooks now live
canonically in @unpunnyfuns/swatchbook-blocks; the addon re-exports the
same names for back-compat with existing imports from
@unpunnyfuns/swatchbook-addon and @unpunnyfuns/swatchbook-addon/hooks.

Closes #202.
