+++ packages/addon/src/hooks/index.ts
@@ -4,4 +4,9 @@ export {
   type TokenInfo,
   type TokenPath,
 } from '#/hooks/use-token.ts';
-export { useActiveAxes, useActiveTheme } from '#/theme-context.ts';
+/**
+ * `useActiveTheme` / `useActiveAxes` live canonically in
+ * `@unpunnyfuns/swatchbook-blocks`. Re-exported here for back-compat with
+ * consumers that imported from `@unpunnyfuns/swatchbook-addon/hooks`.
+ */
+export { useActiveAxes, useActiveTheme } from '@unpunnyfuns/swatchbook-blocks';
