+++ packages/addon/src/index.ts
@@ -8,18 +8,28 @@ export {
   PARAM_KEY,
   VIRTUAL_MODULE_ID,
 } from '#/constants.ts';
-export { AxesContext, ThemeContext, useActiveAxes, useActiveTheme } from '#/theme-context.ts';
 export type { AddonOptions } from '#/options.ts';
+/**
+ * Provider, contexts, and hooks live canonically in
+ * `@unpunnyfuns/swatchbook-blocks`. They are re-exported here for
+ * back-compat with consumers that imported them from the addon prior to
+ * the dep-flip; prefer importing from `@unpunnyfuns/swatchbook-blocks`
+ * going forward.
+ */
 export {
+  AxesContext,
   type ProjectSnapshot,
   SwatchbookContext,
+  ThemeContext,
+  useActiveAxes,
+  useActiveTheme,
   useOptionalSwatchbookData,
   type VirtualAxisShape,
   type VirtualDiagnosticShape,
   type VirtualPresetShape,
   type VirtualThemeShape,
   type VirtualTokenShape,
-} from '#/swatchbook-context.ts';
+} from '@unpunnyfuns/swatchbook-blocks';
 
 /**
  * CSF Next factory. Consumers call this inside
